Splitting vision samples and added region tags.
