+++ vision/spec/detect_landmarks_spec.rb
@@ -1,0 +1,44 @@
+# Copyright 2016 Google, Inc
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+require_relative "../detect_landmarks"
+require "rspec"
+require "tempfile"
+
+describe "Detect Landmarks Sample" do
+
+  # Returns full path to sample image included in repository for testing
+  def image_path filename
+    File.expand_path "../images/#{filename}", __dir__
+  end
+
+  # Capture and return STDOUT output by block
+  def capture &block
+    real_stdout = $stdout
+    $stdout = StringIO.new
+    block.call
+    @captured_output = $stdout.string
+  ensure
+    $stdout = real_stdout
+  end
+  attr_reader :captured_output
+
+  example "detect Eiffel Tower" do
+    expect {
+      detect_landmarks path_to_image_file: image_path("eiffel_tower.jpg")
+    }.to output(
+      "Found landmark: Eiffel Tower\n"
+    ).to_stdout
+  end
+end
